+++ hpOneView/resources/storage/storage_volume_attachments.py
@@ -25,10 +25,10 @@ from future import standard_library
 standard_library.install_aliases()
 
 
-from hpOneView.resources.resource import ResourceClient
+from hpOneView.resources.resource import Resource
 
 
-class StorageVolumeAttachments(object):
+class StorageVolumeAttachments(Resource):
     """
     Storage Volume Attachments API client.
 
@@ -36,33 +36,8 @@ class StorageVolumeAttachments(object):
 
     URI = '/rest/storage-volume-attachments'
 
-    def __init__(self, con):
-        self._connection = con
-        self._client = ResourceClient(con, self.URI)
-
-    def get_all(self, start=0, count=-1, filter='', sort=''):
-        """
-        Gets a list of volume attachment resources.
-
-        Args:
-            start:
-                The first item to return, using 0-based indexing.
-                If not specified, the default is 0 - start with the first available item.
-            count:
-                The number of resources to return. A count of -1 requests all items.
-                The actual number of items in the response might differ from the requested
-                count if the sum of start and count exceeds the total number of items.
-            filter (list or str):
-                A general filter/query string to narrow the list of items returned. The
-                default is no filter; all resources are returned.
-            sort:
-                The sort order of the returned data set. By default, the sort order is based
-                on create time with the oldest entry first.
-
-        Returns:
-            list: Volume attachment resources.
-        """
-        return self._client.get_all(start, count, filter=filter, sort=sort)
+    def __init__(self, connection, data=None):
+        super(StorageVolumeAttachments, self).__init__(connection, data)
 
     def get_extra_unmanaged_storage_volumes(self, start=0, count=-1, filter='', sort=''):
         """
@@ -87,7 +62,7 @@ class StorageVolumeAttachments(object):
             list: Extra unmanaged storage volumes.
         """
         uri = self.URI + "/repair?alertFixType=ExtraUnmanagedStorageVolumes"
-        return self._client.get_all(start=start, count=count, filter=filter, sort=sort, uri=uri)
+        return self._helper.get_all(start=start, count=count, filter=filter, sort=sort, uri=uri)
 
     def remove_extra_presentations(self, resource, timeout=-1):
         """
@@ -104,53 +79,21 @@ class StorageVolumeAttachments(object):
         """
         uri = self.URI + "/repair"
         custom_headers = {'Accept-Language': 'en_US'}
-        return self._client.create(resource, uri=uri, timeout=timeout, custom_headers=custom_headers)
+        return self._helper.create(resource, uri=uri, timeout=timeout, custom_headers=custom_headers)
 
-    def get_paths(self, id_or_uri, path_id_or_uri=''):
+    def get_paths(self, path_id=''):
         """
         Gets all paths or a specific attachment path for the specified volume attachment.
 
         Args:
-            id_or_uri: Can be either the volume attachment id or the volume attachment uri.
-            path_id_or_uri: Can be either the path id or the path uri.
+            path_id: path id
 
         Returns:
             dict: Paths.
         """
-        if path_id_or_uri:
-            uri = self._client.build_uri(path_id_or_uri)
-            if "/paths" not in uri:
-                uri = self._client.build_uri(
-                    id_or_uri) + "/paths" + "/" + path_id_or_uri
-
+        if path_id:
+            uri = "{}/paths/{}".format(self.data["uri"], path_id)
         else:
-            uri = self._client.build_uri(id_or_uri) + "/paths"
+            uri = "{}/paths".format(self.data["uri"])
 
-        return self._client.get(uri)
-
-    def get(self, id_or_uri):
-        """
-        Gets a volume attachment by ID or URI.
-
-        Args:
-            id_or_uri: Can be either the volume attachment ID or the volume attachment URI.
-
-        Returns:
-            dict: volume attachment
-        """
-        return self._client.get(id_or_uri)
-
-    def get_by(self, field, value):
-        """
-        Gets all storage systems that match the filter.
-
-        The search is case-insensitive.
-
-        Args:
-            field: Field name to filter.
-            value: Value to filter.
-
-        Returns:
-            list: List of volume attachments.
-        """
-        return self._client.get_by(field, value)
+        return self._helper.do_get(uri)